Implement home screen UI and some refactoring

+++ app/src/main/java/com/silosoft/technologies/dvtweatherapp/domain/ui_model/ForecastUiModel.kt
@@ -7,5 +7,5 @@ data class ForecastUiModel(
 data class ForecastDay(
     val dayOfWeek: String,
     val weatherType: String,
-    val temp: Double
+    val temp: Int
 )
